trategies have come true. 2017-08-5

+++ driller-tools/get_trace.py
@@ -13,7 +13,7 @@ import trace
 
 l = logging.getLogger("tracer.Tracer")
 
-def dynamic_trace(tracer_qemu,input_path,target_binary,output_dir,test_from_dir,input_from,add_env):
+def dynamic_trace(tracer_qemu,input_path,target_binary,output_dir,test_from,input_from,add_env):
         '''
         record the executed BBs of a testcase
         @param input_from: read from file or stdin 
@@ -38,7 +38,6 @@ def dynamic_trace(tracer_qemu,input_path,target_binary,output_dir,test_from_dir,
             p = subprocess.Popen(
                     args,
                     stdin=subprocess.PIPE,
-                    #stdout=stdout_f,
                     stdout=subprocess.PIPE,
                     stderr=devnull,
                     env=add_env
@@ -46,9 +45,9 @@ def dynamic_trace(tracer_qemu,input_path,target_binary,output_dir,test_from_dir,
             #如果是stdin程序 
             if input_from=="stdin":
                 f=open(input_path, 'rb')
-                input=f.read()
+                input_stdin=f.read()
                 f.close()
-                _,_= p.communicate(input)#读取测试用例,输入 加'\n'后可以多次
+                _,_= p.communicate(input_stdin)#读取测试用例,输入 加'\n'后可以多次
                 
             ret = p.wait() #等待返回结果
             
@@ -59,6 +58,7 @@ def dynamic_trace(tracer_qemu,input_path,target_binary,output_dir,test_from_dir,
                             during dynamic tracing", abs(ret))
                     l.info("entering crash mode")
                     is_crash_case =True #表示这是一个crash测试用例
+                    print input_path
 
             stdout_f.close()
         # end 得到一个轨迹
@@ -75,6 +75,8 @@ def dynamic_trace(tracer_qemu,input_path,target_binary,output_dir,test_from_dir,
 #                  if v.startswith('Trace')] # 得到所有的基本块地址,删掉了别的内容 str类型
         
         addrs = [v.split('[')[1] for v in trace.split('\n') if v.startswith('Trace')] # 得到所有的基本块地址,这里保留函数名称 str类型
+        addrs_set=set()
+        addrs_set.update(addrs)  # 去掉重复的轨迹
         
         # grab the faulting address
         if is_crash_case:
@@ -82,51 +84,50 @@ def dynamic_trace(tracer_qemu,input_path,target_binary,output_dir,test_from_dir,
             #print trace
             #print trace.split('\n')[-2]
             #print trace.split('\n')[-1]#这个是空格
-            crash_addr = [trace.split('\n')[-2].split('[')[1]]         #最后一个基本块 address
-        
+            crash_addr = [trace.split('\n')[-2].split('[')[1]]         #最后一个基本块 address 奔溃点的地址
         
         #输出每个测试用例的轨迹
         #配置每个测试用例的输出名称
-        input_name=os.path.basename(input_path)
+        input_name =os.path.basename(input_path)
         input_name = 'id'+input_name.split("id:")[-1].split(",")[0]
         
-        #应该根据来源保存
-        if 0:
+        #应该根据来源保存,如果是crash,则添加名称
+        if 1:
             if is_crash_case:
                 input_name+='crashes'
-            else:
-                input_name+='queue'
-        #根据执行结果保存  
-        if 1:  
+#             else:
+#                 input_name+='queue'
+        
+        #根据执行结果保存,修改保存目录
+        if 0:  
             if is_crash_case:
-                test_from_dir='crashes'
+                test_from='crash'
             else:
-                test_from_dir='queue'    
-        addrs_set=set()
-        addrs_set.update(addrs)  # 去掉重复的
-        write_each_trace(output_dir,input_name, addrs, addrs_set, test_from_dir)
+                test_from='queue'    
+        
+        write_each_trace(output_dir,input_name, addrs, addrs_set, test_from)
             
         os.remove(lname)#删除记录测试用例轨迹的临时文件
         return (addrs,crash_addr,addrs_set)  #返回一个list  如果是crash,addrs是包括最后一个的
 
 def write_out_all_trace(trace_set,crash_set,output_dir,trace_num_dict,trace_num_dict_set): #包括crash和queue
     #输出目录
-    out_trace = os.path.join(output_dir, "trace_all")
-    out_crash= os.path.join(output_dir, "crashes_blocks")  #只记录崩溃处的位置
+    add_trace = os.path.join(output_dir, "trace_all")
+    crash_addrs= os.path.join(output_dir, "all_crashes_addrs")  #只记录崩溃处的位置
     trace_num= os.path.join(output_dir, "trace_num")
     trace_num_set= os.path.join(output_dir, "trace_num_set")
     
-    if os.path.exists(out_trace):
-        os.remove(out_trace)
-    if os.path.exists(out_crash):
-        os.remove(out_crash)  
+    if os.path.exists(add_trace):
+        os.remove(add_trace)
+    if os.path.exists(crash_addrs):
+        os.remove(crash_addrs)  
     if os.path.exists(trace_num):
         os.remove(trace_num)  
     if os.path.exists(trace_num_set):
         os.remove(trace_num_set)     
     
     #输出所有轨迹地址集合 (16进制)
-    with open(out_trace+"hex", 'a') as ofp:
+    with open(add_trace+"hex", 'a') as ofp:
         for v in trace_set:
             a=v.split(']')[0]
             b=v.split(']')[1]
@@ -134,7 +135,7 @@ def write_out_all_trace(trace_set,crash_set,output_dir,trace_num_dict,trace_num_
             ofp.write('\n')
                 
     #输出所有轨迹地址集合 (10进制)
-    with open(out_trace, 'a') as ofp:
+    with open(add_trace, 'a') as ofp:
         for v in trace_set:
             a=v.split(']')[0]
             b=v.split(']')[1]
@@ -143,8 +144,10 @@ def write_out_all_trace(trace_set,crash_set,output_dir,trace_num_dict,trace_num_
             ofp.write(a+b)  # 将内容输出到目标目录
             ofp.write('\n')
     #---------------------------------------------
+    
+    
     #输出崩溃点地址集合 (16进制)
-    with open(out_crash+"hex", 'a') as ofp:
+    with open(crash_addrs+"hex", 'a') as ofp:
         for v in crash_set:
             a=v.split(']')[0]
             b=v.split(']')[1]
@@ -152,7 +155,7 @@ def write_out_all_trace(trace_set,crash_set,output_dir,trace_num_dict,trace_num_
             ofp.write('\n')
                 
     #输出崩溃点地址集合  (10进制)
-    with open(out_crash, 'a') as ofp:
+    with open(crash_addrs, 'a') as ofp:
         for v in crash_set:
             a=v.split(']')[0]
             b=v.split(']')[1]
@@ -160,7 +163,10 @@ def write_out_all_trace(trace_set,crash_set,output_dir,trace_num_dict,trace_num_
             a=str(a)
             ofp.write(a+b)  # 将内容输出到目标目录
             ofp.write('\n')  
-            
+    #  -----------------------------------------------------------------------------
+    
+    
+    
     # 输出基本块数量, 包含重复的
     with open(trace_num, 'a') as ofp:
         ofp.write("all_num is(no duplicate) :"+str(len(trace_set)) )
@@ -169,6 +175,8 @@ def write_out_all_trace(trace_set,crash_set,output_dir,trace_num_dict,trace_num_
             a=k+":"+str(v)
             ofp.write(a)  # 将内容输出到目标目录
             ofp.write('\n')    
+    #  -----------------------------------------------------------------------------
+    
     
     # 输出基本块数量, 不包含重复的
     with open(trace_num_set, 'a') as ofp:
@@ -179,22 +187,22 @@ def write_out_all_trace(trace_set,crash_set,output_dir,trace_num_dict,trace_num_
             ofp.write(a)  # 将内容输出到目标目录
             ofp.write('\n')             
 
-def write_each_trace(output_dir,input_name, test_trace, test_trace_set,from_dir):
+def write_each_trace(output_dir,input_name, test_trace, test_trace_set,test_from):
     
     '''
     @param output_dir: the output directory
     @param input_name:  the id of the test_case
     @param test_trace: the trace of the test_case 
-    @param from_dir: indicate what is the test-case
+    @param test_from: indicate what is the test-case
     '''
     
     #选择输出目录
-    if from_dir=='queue':
+    if test_from=='queue':
         out_trace = os.path.join(output_dir, "queue")
-    elif from_dir=='crashes':
+    elif test_from=='crash':
         out_trace = os.path.join(output_dir, "crashes")
     else:
-        print"no queue or crashes"
+        print"error test_from"
         exit(1)  
     
     test_trace_set=set()
@@ -223,51 +231,38 @@ def write_each_trace(output_dir,input_name, test_trace, test_trace_set,from_dir)
             
             
                  
-def start_get_trace(target_binary):
-    
-    #配置目标程序
-    #target_binary = "/home/xiaosatianyu/Desktop/driller/binary-cgc/YAN01_00016" #这个可以读取 bmp2tiff claw32
-    #target_binary = "/home/xiaosatianyu/Desktop/afl-yyy/target/brancher" #这个可以读取 bmp2tiff claw32
-    
+def start_get_trace(binary_path,data_dir):
     #配置对应的qemu
-#     qemu_dir="/home/xiaosatianyu/workspace/git/driller-yyy/shellphish-qemu/shellphish_qemu/bin" #来自于tracer
-#     #p = angr.Project(target_binary)
-#     #platform = p.arch.qemu_name
-#     platform = 'cgc'
-#     if platform == 'i386':
-#         tracer_qemu = os.path.join(qemu_dir, "shellphish-qemu-linux-i386")
-#     elif platform == 'x86_64': 
-#         tracer_qemu = os.path.join(qemu_dir, "shellphish-qemu-linux-x86_64")
-#     elif platform == 'cgc': 
-#         tracer_qemu = os.path.join(qemu_dir, "shellphish-qemu-cgc-tracer")
-#     else:
-#         print "no qemu\n"
-#         exit(1)     
+    qemu_dir="/home/xiaosatianyu/workspace/git/driller-yyy/shellphish-qemu/shellphish_qemu/bin" #来自于tracer
+    p = angr.Project(binary_path)
+    platform = p.arch.qemu_name
+    if platform == 'i386':
+        tracer_qemu = os.path.join(qemu_dir, "shellphish-qemu-linux-i386")
+    elif platform == 'x86_64': 
+        tracer_qemu = os.path.join(qemu_dir, "shellphish-qemu-linux-x86_64")
+    elif platform == 'cgc': 
+        tracer_qemu = os.path.join(qemu_dir, "shellphish-qemu-cgc-tracer")
+    else:
+        print "no qemu\n"
+        exit(1)     
         
-    #配置afl-cgc的qemu
-    #afl-cgc下的qemu
-    qemu_dir_cgc="/home/xiaosatianyu/workspace/git/driller-yyy/shellphish-afl/bin/afl-cgc/tracers/i386"
-    tracer_qemu = os.path.join(qemu_dir_cgc, "afl-qemu-trace")
-    
     #配置测试用例输入目录,这个是当个afl引擎的情况
     #input_from ="file" 
     input_from = "stdin"
     
-    afl_dir="/tmp/driller/file/sync/fuzzer-master"
-    #test_case_dir = os.path.join(afl_dir, "queue")  # AFL生成测试用例的目录
-    test_case_dir = "/home/xiaosatianyu/Desktop/driller-desk/seed"
-    #test_case_dir = "/tmp/driller/YAN01_00016/driller"
-#     test_case_dir=os.path.join("/tmp/driller",os.path.basename(target_binary),"driller/queue")
+    queue_dir = os.path.join(data_dir, "queue")  # AFL生成测试用例的目录
+    #queue_dir = "/home/xiaosatianyu/Desktop/driller-desk/seed"
+    #queue_dir=os.path.join("/tmp/driller",os.path.basename(binary_path),"driller/queue")
     
-    crash_dir = os.path.join(afl_dir, "crashes")  # AFL生成测试用例的目录
+    crash_dir = os.path.join(data_dir, "crashes")  # AFL生成测试用例的目录
     
-    #配置输出目录
-    target_base=os.path.basename(target_binary) #在tmp/程序名 下
-    output_dir=os.path.join("/tmp/traces",target_base)
+    if not os.path.exists(queue_dir) and not os.path.exists(crash_dir):
+        return 0
+    
+    #配置输出目录  #创建目录
+    output_dir=os.path.join("/tmp/traces",os.path.basename(binary_path))
     if os.path.isdir(output_dir):
         shutil.rmtree(output_dir) #删除工作目录
-    
-    #创建目录
     os.makedirs(output_dir)
     os.makedirs(os.path.join(output_dir,"queue"))  #保存
     os.makedirs(os.path.join(output_dir,"crashes"))
@@ -275,7 +270,7 @@ def start_get_trace(target_binary):
     #配置环境变量
     add_env={"HOME": os.environ["HOME"]}   
     
-    #完成配置----
+    # 完成配置-------------- ------------------------------------------------------
     
     trace_set=set() #记录所有的基本块,包含queue和crash下的
     crash_block_set=set() #记录所有崩溃的最后一个基本块
@@ -283,8 +278,8 @@ def start_get_trace(target_binary):
     trace_num_dict_set=dict() #记录基本块的数量 不包含重复的
     
     #read test-cases
-    if os.path.isdir(test_case_dir):
-        queue_inputs = filter(lambda d: not d.startswith('.'), os.listdir(test_case_dir))  # queue下的测试用例
+    if os.path.isdir(queue_dir):
+        queue_inputs = filter(lambda d: not d.startswith('.'), os.listdir(queue_dir))  # queue下的测试用例
     else:
         queue_inputs=[]   
         
@@ -294,12 +289,13 @@ def start_get_trace(target_binary):
     else:
         crash_inputs=[]
         
-    #遍历queue目录. 得到测试用例的执行轨迹 driller的queue目录下也会有crash
-    for input_file in queue_inputs:
-        input_data_path = os.path.join(test_case_dir, input_file) 
-        test_from_dir="queue" # 也包括任意目录下的测试用例
-        addrs,crash_address,addrs_set=dynamic_trace(tracer_qemu,input_data_path,target_binary,output_dir,test_from_dir, input_from, add_env=add_env)#记录对应测试用例的轨迹
+    #遍历queue目录. 得到测试用例的执行轨迹; 如果是driller的queue目录下也会有crash 
+    for queue_input in queue_inputs:
+        input_data_path = os.path.join(queue_dir, queue_input) 
+        test_from="queue" # 也包括任意目录下的测试用例
+        addrs,crash_address,addrs_set=dynamic_trace(tracer_qemu,input_data_path,binary_path,output_dir,test_from, input_from, add_env=add_env)#记录对应测试用例的轨迹
         
+        #得到轨迹
         # 这里可以使用set得到不重复的 trace
         trace_set.update(addrs)
         crash_block_set.update(crash_address) #记录的是崩溃处的地址  
@@ -307,10 +303,12 @@ def start_get_trace(target_binary):
         trace_num_dict_set.update({os.path.basename(input_data_path):len(addrs_set)})
         
     #遍历crash目录. 得到测试用例的执行轨迹
-    for input_file in crash_inputs:
-        input_data_path = os.path.join(crash_dir, input_file) 
-        test_from_dir="crash"
-        addrs,crash_address,_=dynamic_trace(tracer_qemu,input_data_path,target_binary,output_dir,test_from_dir,input_from,add_env=add_env)#记录对应测试用例的轨迹
+    for crash_input in crash_inputs:
+        input_data_path = os.path.join(crash_dir, crash_input) 
+        test_from="crash"
+        addrs,crash_address,_=dynamic_trace(tracer_qemu,input_data_path,binary_path,output_dir,test_from,input_from,add_env=add_env)#记录对应测试用例的轨迹
+        
+        #得到轨迹
         #这里可以使用set得到不重复的 trace
         trace_set.update(addrs)    
         crash_block_set.update(crash_address) #记录的是崩溃处的地址 
@@ -318,28 +316,27 @@ def start_get_trace(target_binary):
     #输出所有的轨迹到文件
     write_out_all_trace(trace_set,crash_block_set,output_dir,trace_num_dict,trace_num_dict_set)
     
-   
-    
-    return len(queue_inputs) #返回符号执行生成的数量
-    
+    return len(queue_input)#返回测试用例的数量
+
+
 if __name__ == "__main__":
     print "start!\n"
-    binaries_dir="/home/xiaosatianyu/Desktop/driller-desk/binary-cgc-test"
+    binaries_dir="/home/xiaosatianyu/Desktop/driller-desk/binary-cgc"
+    binaries_dir="/home/xiaosatianyu/CTF/AFL-execution/target"
     target_pros=os.listdir(binaries_dir)
     target_pros.sort()
     if os.path.exists("/tmp/traces"):
         shutil.rmtree("/tmp/traces")
     os.makedirs("/tmp/traces")
-    sym_num_path="/tmp/traces/num_of_symbolic_gen"
-    if os.path.exists(sym_num_path):
-        os.remove(sym_num_path)
     
-    ofp=open(sym_num_path, 'a')    
-    for i in target_pros:
-        target_pro=os.path.join(binaries_dir,i)
-        num=start_get_trace(target_pro) #收集符号执行产生的数量
-        ofp.write(os.path.basename(target_pro)+" generate "+str(num)+"\n")
-        ofp.flush()
-        print "get %s ok" % os.path.basename(target_pro)
+    
+    #data_dir= os.path.join("/tmp/driller",os.path.basename(binary_path),"sync/fuzzer-master")
+    
+    for pro in target_pros:
+        binary_path=os.path.join(binaries_dir,pro)
+        data_dir="/tmp/sync-ctf/1"
+        num=start_get_trace(binary_path,data_dir) #收集符号执行产生的数量
+        if num!=0:
+            print "get %s ok" % os.path.basename(binary_path)
     print "end!\n"
     sys.exit()
